OAK-4131: LastRevisionRecoveryAgent may throw ClassCastException
Implement fix and enable test

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/branches/1.0@1735494 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-core/src/main/java/org/apache/jackrabbit/oak/plugins/document/mongo/MongoMissingLastRevSeeker.java
@@ -86,7 +86,7 @@ public class MongoMissingLastRevSeeker extends MissingLastRevSeeker {
         BasicDBObject setUpdates = new BasicDBObject();
         setUpdates.append(ClusterNodeInfo.REV_RECOVERY_LOCK, RecoverLockState.ACQUIRED.name());
         if (recoveredBy != 0) {
-            setUpdates.append(ClusterNodeInfo.REV_RECOVERY_BY, recoveredBy);
+            setUpdates.append(ClusterNodeInfo.REV_RECOVERY_BY, (long) recoveredBy);
         }
 
         BasicDBObject update = new BasicDBObject();
